52done

+++ routes/campgrounds.js
@@ -1,9 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const Campground = require('../models/campground');
-const { isLoggedIn } = require('../middleware');
+const { isLoggedIn , isAuthor,validateCampground} = require('../middleware');
 
-const { campgroundSchema } = require('../schemas.js');
 
 const catchAsync = require('../utils/catchAsync');
 const ExpressError = require('../utils/ExpressError');
@@ -23,24 +22,11 @@ router.get('/new', isLoggedIn,(req, res) => {
 });
 
 
-
-const validateCampground = (req, res, next) => {
-
-    const { error } = campgroundSchema.validate(req.body);
-    if (error) {
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400)
-    } else {
-        next();
-    }
-}
-
-
-
-
 router.post('/',isLoggedIn ,validateCampground, catchAsync(async (req, res, next) => {
     // if (!req.body.campground) throw new ExpressError('Invalid Campground Data', 400);
+
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id; // Set the author to the logged-in user
     await campground.save();
     req.flash('success', 'Successfully made a new campground!');
     res.redirect(`/campgrounds/${campground._id}`)
@@ -49,7 +35,8 @@ router.post('/',isLoggedIn ,validateCampground, catchAsync(async (req, res, next
 
 router.get('/:id',catchAsync(async (req, res) => {
 
-    const campground = await Campground.findById(req.params.id).populate('reviews');
+
+    const campground = await Campground.findById(req.params.id).populate({path :'reviews', populate :{path : 'author'}}).populate('author');
 
     if(!campground) {
         req.flash('error', 'Cannot find that campground!');
@@ -60,18 +47,23 @@ router.get('/:id',catchAsync(async (req, res) => {
 
 }));
 
-router.get('/:id/edit', isLoggedIn,catchAsync(async (req, res) => {
-
-        const campground = await Campground.findById(req.params.id);
+router.get('/:id/edit', isLoggedIn,isAuthor,catchAsync(async (req, res) => {
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if(!campground) {   
+        req.flash('error', 'Cannot find that campground!');
+        return res.redirect('/campgrounds');
+    }
+    
         res.render('campgrounds/edit',{campground});
 
 
 }));
 
 
-router.put('/:id',isLoggedIn,validateCampground, catchAsync(async (req, res) => {
+router.put('/:id',isLoggedIn,isAuthor,validateCampground, catchAsync(async (req, res) => {
 
-    const { id } = req.params;
+
     const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash('success', 'Successfully updated campground!');
     res.redirect(`/campgrounds/${campground._id}`);
@@ -80,9 +72,16 @@ router.put('/:id',isLoggedIn,validateCampground, catchAsync(async (req, res) => 
 
 }));
 
-router.delete('/:id', catchAsync(async (req, res) => {
+router.delete('/:id',isLoggedIn,isAuthor, catchAsync(async (req, res) => {
 
     const { id } = req.params;
+    
+        const campground = await Campground.findById(id);
+    if(!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission to edit this campground');
+        return res.redirect(`/campgrounds/${id}`);
+    }
+    
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted campground!');
     res.redirect('/campgrounds');
